Polished source code, established generic identifiers

+++ portfolioBack/src/main/java/com/yoprogramo/portfolioBack/model/Persona.java
@@ -1,3 +1,4 @@
+
 package com.yoprogramo.portfolioBack.model;
 
 import javax.persistence.Entity;
@@ -9,28 +10,24 @@ import javax.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter
-@Setter
+@Getter @Setter
 @Entity
 public class Persona {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private int id;
     @NotNull
     @Size(min = 1, max = 50, message = "Length not allowed")
     private String nombre;
-
     @NotNull
     @Size(min = 1, max = 50, message = "Length not allowed")
     private String apellido;
 
     public Persona() {
-
     }
 
-    public Persona(Long id, String nombre, String apellido) {
+    public Persona(int id, String nombre, String apellido) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
