+++ portfolioBack/src/main/java/com/yoprogramo/portfolioBack/repository/IRPersona.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRPersona extends JpaRepository <Persona, Long> {
+public interface IRPersona extends JpaRepository <Persona, Integer> {
     
 }
